# Fast addition in cpp_int