Compilers currently generate *bad* assembly so, with a little help to the compiler, I was able to reduce the number of instructions (`mov` and `add` both) in the addition loop. 

[See the Godbolt link showing the difference in assembly](https://godbolt.org/z/lH1_Hc)

Benefits:
- Can be easily unrolled
- Easy to read
- Efficient (?)

```
// clang 10.0, Arch Linux, -std=c++11
// Current develop
BM<bst_int>/40000        3826 ns         3730 ns       191511
BM<bst_int>/80000        4545 ns         4486 ns       157061
BM<bst_int>/120000       5416 ns         5335 ns       135902
BM<bst_int>/160000       6238 ns         6139 ns       113651
BM<bst_int>/200000       6924 ns         6829 ns       103996

// This PR
BM<bst_int>/40000        3395 ns         3362 ns       208408
BM<bst_int>/80000        3954 ns         3921 ns       172037
BM<bst_int>/120000       4551 ns         4511 ns       158986
BM<bst_int>/160000       5120 ns         5063 ns       136567
BM<bst_int>/200000       5545 ns         5496 ns       112399

BM<mpz_int>/40000         751 ns          744 ns       961070
BM<mpz_int>/80000        1536 ns         1522 ns       461803
BM<mpz_int>/120000       2389 ns         2371 ns       297695
BM<mpz_int>/160000       3475 ns         3439 ns       215895
BM<mpz_int>/200000       4008 ns         3972 ns       178096
```

Concerns:
- Most importantly will boost take this path of using intrinsics?
- Still unable to see huge performance gains as `mov` instructions are reduced and also `add` is executed only once instead of 4 times. Needs some investigation.
- GCC cannot unroll the loop while clang can, thus there is a huge disparity between GCC & LLVM.
- This assembly can be optimized further by writing assembly by hand (inline asm?) but still has potential.
- Portability [New!]