@NAThompson 2020-05-03
> Most importantly will boost take this path of using intrinsics?

Is there a macro that can detect the presence of the intrinsic? If so it looks like a pretty unambiguous win.