@ckormanyos 2020-05-03
Have you tried to change the original bottleneck line to:

`carry += static_cast<double_limb_type>(*pa) + *pb;`

It looks like the disassembled code might be performing ful (double-limb + double-limb) add, which would in fact be slow.

I did not chek this. But you might consider to do so.

Kind regards, Chris