@ckormanyos 2020-05-03
You could even try:
```
carry += *pa;
carry += *pb;
```
... to get oriented with how these particular compilers might _think_, which I also did not actually do. So just an idea here...