@madhur4127 2020-05-03
> carry += *pa;
> carry += *pb;

This should take 4 64 bit add instructions to execute. My version generates only 1  without the need for double limb.

> BTW how is mpz_int so fast? What does its asm look like?

LLVM with support from Intel recently announced _ExtInt which generates most efficient assembly, you may look at that.

Still few instructions can be optimised if I could use ADCX instruction.