@jzmaddock 2020-05-03
@madhur4127 the compiler intrinsic corresponding to ADDCX is:

```
unsigned char _addcarryx_u64 (unsigned char c_in, unsigned __int64 src1, unsigned __int64 src2, unsigned __int64 *sum_out);
```

@NAThompson : yes we can configure this, though it will take a bit of headscratching ;)  I can take care of that if @madhur4127 can optimise this as far as it's likely to go.