@madhur4127 2020-05-04
> @madhur4127 the compiler intrinsic corresponding to ADDCX is:

This cannot be made to work without ask. Because:
1. GCC doesn't recognize CF and OF carry flag so ADOX, ADCX cannot be made to work at all. 
2. It's hard to fight the compiler to generate the assembly that I want, it is not possible at all.
3. Even though ICC supports it, the generated assembly is ridiculous.

So the Intel's ADX is out of the equation. Its sad to see Intel adding two instructions to support big integer arithmetic without making it *easy* enough to use it.

>  yes we can configure this, though it will take a bit of headscratching ;) I can take care of that if @madhur4127 can optimise this as far as it's likely to go.
 
I am able to generate an almost optimal assembly with Clang with negligible overhead (1  per 22 instructions in the loop). This the best that can be done. [See Godbolt for assembly](https://godbolt.org/z/tCDVy9).

This PR will use `ADC` which has been part of ISA since the times of 8086 and is present in every processor and architecture. So no head-scratching I guess ([See Intel's intrinsic guide](https://software.intel.com/sites/landingpage/IntrinsicsGuide/#text=_addcarry_u64)).

Here is the comparison with GMP. There are many more things at play here that might have caused the difference and I am confident (now) that addition is not responsible for it.

GCC cannot optimize the way that Clang did, this is a missed optimization bug which I will file soon. It generated bad assembly before and now it has improved (but still is too slow as compared to clang).


```
// Current develop
BM<bst_int>/40000        3826 ns         3730 ns       191511
BM<bst_int>/80000        4545 ns         4486 ns       157061
BM<bst_int>/120000       5416 ns         5335 ns       135902
BM<bst_int>/160000       6238 ns         6139 ns       113651
BM<bst_int>/200000       6924 ns         6829 ns       103996

// This PR
BM<bst_int>/40000        3395 ns         3362 ns       208408
BM<bst_int>/80000        3954 ns         3921 ns       172037
BM<bst_int>/120000       4551 ns         4511 ns       158986
BM<bst_int>/160000       5120 ns         5063 ns       136567
BM<bst_int>/200000       5545 ns         5496 ns       112399

BM<mpz_int>/40000         751 ns          744 ns       961070
BM<mpz_int>/80000        1536 ns         1522 ns       461803
BM<mpz_int>/120000       2389 ns         2371 ns       297695
BM<mpz_int>/160000       3475 ns         3439 ns       215895
BM<mpz_int>/200000       4008 ns         3972 ns       178096
```