@NAThompson 2020-05-04
@madhur4127 : What are your compiler flags? Looking at your godbolt, you might benefit from `-march=native`, or (on goldbolt) `-march=skylake`.