@madhur4127 2020-05-04
> What are your compiler flags?

`-O3 -march=native`

> Looking at your godbolt, you might benefit from -march=native, or (on goldbolt) -march=skylake.

I have a coffee lake CPU and `-march=native` or `-march=skylake` made no difference. In fact I guess, there shouldn't be any difference as the instructions used are add, adc, lea, mov. Which I guess every processor of this century will support.

Here's the code used in benchmarking:
```c++
#include <iostream>
#include <benchmark/benchmark.h>
#include <boost/multiprecision/cpp_int.hpp>
#include <boost/multiprecision/gmp.hpp>
#include <boost/random.hpp>
#include <cmath>

using namespace boost::multiprecision;
using namespace boost::random;

template <typename T>
static void BM(benchmark::State& state)
{
   mt19937_64                  mt(223);
   int                         bits = state.range(0);
   uniform_int_distribution<T> ui((T(1) << (bits - 1) + 1), T(1) << bits);
   T                           x = ui(mt), y = ui(mt), z;

   for (auto _ : state)
   {
	  benchmark::DoNotOptimize(z = x + y);
	  x++;
   }
}
const int start = 4e4, end = 2e5, step = 4e4;
using bst_int = number<cpp_int_backend<2 * end, 2 * end, unsigned_magnitude, unchecked>, et_off>;
BENCHMARK_TEMPLATE(BM, bst_int)->DenseRange(start, end, step);
BENCHMARK_TEMPLATE(BM, mpz_int)->DenseRange(start, end, step);
BENCHMARK_MAIN();

```