@madhur4127 2020-05-04
The difference between GCC/ICC vs Clang is very much significant.

I think as per performance is concerned, there is an improvement over popular compilers (MSVC?).

```
Flags: -O3 -march=native

// GCC (gcc version 9.3.0 (Arch Linux 9.3.0-1)), Old:
BM<bst_int>/40000        8290 ns         6509 ns        97295
BM<bst_int>/80000        8039 ns         7914 ns        87575
BM<bst_int>/120000      10061 ns         9873 ns        70388
BM<bst_int>/160000      12735 ns        12250 ns        58322
BM<bst_int>/200000      14449 ns        13855 ns        51252

//GCC, New:
BM<bst_int>/40000        5413 ns         5350 ns       117355
BM<bst_int>/80000        6715 ns         6538 ns       104644
BM<bst_int>/120000      10042 ns         7391 ns        91100
BM<bst_int>/160000       8398 ns         8116 ns        79428
BM<bst_int>/200000       8839 ns         8727 ns        80258

// ICC(icc version 19.1.1.219 (gcc version 9.3.0 compatibility)), Old
BM<bst_int>/40000        9426 ns         9330 ns        73675
BM<bst_int>/80000       10277 ns        10204 ns        66700
BM<bst_int>/120000      11353 ns        11282 ns        59955
BM<bst_int>/160000      11947 ns        11863 ns        56598
BM<bst_int>/200000      19063 ns        14264 ns        55008

// ICC, New
BM<bst_int>/40000        8688 ns         8471 ns        64066
BM<bst_int>/80000        9894 ns         9764 ns        73916
BM<bst_int>/120000      10460 ns        10364 ns        67272
BM<bst_int>/160000      11647 ns        11491 ns        61775
BM<bst_int>/200000      12290 ns        12179 ns        57930
```